feat(wallet-dashboard): refactor staking dialogs and introduce Validator component
