+++ apps/wallet-dashboard/components/Dialogs/Staking/index.ts
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export { default as StakeDialog } from './StakeDialog';
+export * from './StakedDetailsDialog';
